HTTP return codes update

+++ api/pwd_reset.php
@@ -1,9 +1,11 @@
 <?php
 session_start();
 require "../dls_db.php";
+require "utils.php";
 $_SESSION["resetPwd"] = null;
-if (isset($_POST["t"]) && isset($_POST["password"])) {
-    if (isset($_POST["recaptcha_token"])) {
+
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    if (isset($_POST["t"]) && isset($_POST["password"]) && isset($_POST["recaptcha_token"])) {
         $url = 'https://www.google.com/recaptcha/api/siteverify';
         $data = array('secret' => $captcha_secret, 'response' => $_POST["recaptcha_token"]);
         $options = array(
@@ -15,39 +17,17 @@ if (isset($_POST["t"]) && isset($_POST["password"])) {
         );
         $context  = stream_context_create($options);
         $result = json_decode(file_get_contents($url, false, $context));
-        
-        if (!empty($_SERVER['HTTP_CLIENT_IP'])) { //check ip from share internet
-            $ip=$_SERVER['HTTP_CLIENT_IP'];
-        } elseif (!empty($_SERVER['HTTP_X_FORWARDED_FOR'])) { //to check ip is pass from proxy
-            $ip=$_SERVER['HTTP_X_FORWARDED_FOR'];
-        } else {
-            $ip=$_SERVER['REMOTE_ADDR'];
-        }
     
         if ($result && $result->success && $result->score > 0.5 && $result->action == "pwd_reset" && $result->hostname == "dls.rw.jachyhm.cz") {
             header('Content-type: application/json');
             $pwd = trim($_POST["password"]);
             if (empty($_POST["t"])) {
-                $response = new stdClass();
-                $response->code = -1;
-                $response->message = "Pasword reset token already expired or not valid!";
-            
-                $response_json = json_encode($response);
-            
-                db_log(13, false, -1, $ip, $_POST["t"], "Password reset token invalid!", $mysqli);
-                $mysqli->close();
-                die($response_json);
+                db_log(13, false, -1, $ip, $_POST["t"], "Password reset token not set!", $mysqli);
+                flushResponse(400, "Pasword reset token must be set!", $mysqli);
             }
             if (empty($pwd)) {
-                $response = new stdClass();
-                $response->code = -1;
-                $response->message = "Pasword can not be empty!";
-            
-                $response_json = json_encode($response);
-            
                 db_log(13, false, -1, $ip, $_POST["t"], "Password is empty!", $mysqli);
-                $mysqli->close();
-                die($response_json);
+                flushResponse(400, "Pasword must be set!", $mysqli);
             }
             $sql = $mysqli->prepare('SELECT * FROM `pwd_resets` WHERE `token` = ?;');
             $sql->bind_param('s', $_POST["t"]);
@@ -65,61 +45,47 @@ if (isset($_POST["t"]) && isset($_POST["password"])) {
                     $sql = $mysqli->prepare('DELETE FROM `pwd_resets` WHERE `token` = ?;');
                     $sql->bind_param('s', $_POST["t"]);
                     $sql->execute();
+
                     db_log(13, true, $row["user_id"], $ip, $_POST["t"], "Password changed!", $mysqli);
+                    flushResponse(200, "Pasword changed successfully!", $mysqli);
                 }
             }
-                    
-            $response = new stdClass();
-            $response->code = 1;
-            $response->message = "Pasword changed successfully!";
 
-            $response_json = json_encode($response);
-
-            $mysqli->close();
-            die($response_json);
+            db_log(13, false, -1, $ip, $_POST["t"], "Password reset token invalid!", $mysqli);
+            flushResponse(404, "Pasword reset token already expired or not valid!", $mysqli);
         } else {
-            $response = new stdClass();
-            $response->code = -3;
-            $response->message = "Sorry, but you seem to be a robot. And we definitelly do not want one here.";
-    
-            $response_json = json_encode($response);
-    
             db_log(13, false, $ip, $_POST["t"], "ReCaptcha failed!", $mysqli);
-            $mysqli->close();
-            die($response_json);
+            flushResponse(403, "Sorry, but you seem to be a robot. And we definitelly do not want one here.", $mysqli);
         }
     } else {
-        $response = new stdClass();
-        $response->code = -1;
-        $response->message = "Missing required parameter!";
-    
-        $response_json = json_encode($response);
-    
-        $mysqli->close();
-        die($response_json);
+        flushResponse(400, "Not all parameters set!", $mysqli);
     }
-} elseif (isset($_GET["t"])) {
-    date_default_timezone_set("Europe/Prague");
-    $timestamp = date("Y-m-d H:i:s", time()-3600*48);
-    $sql = $mysqli->prepare('DELETE FROM `pwd_resets` WHERE `generated` < ?;');
-    $sql->bind_param('s', $timestamp);
-    $sql->execute();
+} elseif ($_SERVER['REQUEST_METHOD'] === 'GET') { 
+    if (isset($_GET["t"])) {
+        date_default_timezone_set("Europe/Prague");
+        $timestamp = date("Y-m-d H:i:s", time()-3600*48);
+        $sql = $mysqli->prepare('DELETE FROM `pwd_resets` WHERE `generated` < ?;');
+        $sql->bind_param('s', $timestamp);
+        $sql->execute();
 
-    $sql = $mysqli->prepare('SELECT * FROM `pwd_resets` WHERE `token` = ?;');
-    $sql->bind_param('s', $_GET["t"]);
-    $sql->execute();
-    $queryResult = $sql->get_result();
+        $sql = $mysqli->prepare('SELECT * FROM `pwd_resets` WHERE `token` = ?;');
+        $sql->bind_param('s', $_GET["t"]);
+        $sql->execute();
+        $queryResult = $sql->get_result();
 
-    if (!empty($queryResult)) {
-        if ($queryResult->num_rows > 0) {
-            $mysqli->close();
-            $_SESSION["resetPwd"] = $_GET["t"];
-            header("Location: /");
-            die();
+        if (!empty($queryResult)) {
+            if ($queryResult->num_rows > 0) {
+                $mysqli->close();
+                $_SESSION["resetPwd"] = $_GET["t"];
+                header("Location: /");
+                die();
+            }
         }
     }
+    $_SESSION["errorMessage"] = "Pasword reset token already expired or not valid!";
+    header("Location: /");
+    die();
+} else {
+    flushResponse(405, "Protocol not supported!", $mysqli);
 }
-$_SESSION["errorMessage"] = "Pasword reset token already expired or not valid!";
-header("Location: /");
-die();
 ?>
